Separate files into stakeholder folders
