--- frontend/src/components/AdminStakeholder.scss
@@ -1,1 +1,0 @@
-@import './../../node_modules/react-tabs/style/react-tabs.scss';
