+++ frontend/src/components/PFCAdmin/AdminStakeholder.scss
@@ -1,0 +1,1 @@
+@import './../../../node_modules/react-tabs/style/react-tabs.scss';
