+++ frontend/src/index.js
@@ -7,13 +7,13 @@ import registerServiceWorker from './registerServiceWorker';
 import rootReducer from './reducers';
 import LandingPage from './components/LandingPage';
 import LoginPage from './components/LoginContainer';
-import NewProject from './components/NewProject';
-import ProjectPage from './components/ProjectPage';
-import AdminDashboard from './components/AdminDashboard';
-import AdminStakeholder from './components/AdminStakeholder';
-import AdminNewStakeholder from './components/AdminNewStakeholder';
-import CreateWastePicker from './components/CreateWastePicker';
-import CreateExternalDWCC from './components/CreateExternalDWCC';
+import NewProject from './components/PFCAdmin/NewProject';
+import ProjectPage from './components/PFCAdmin/ProjectPage';
+import AdminDashboard from './components/PFCAdmin/AdminDashboard';
+import AdminStakeholder from './components/PFCAdmin/AdminStakeholder';
+import AdminNewStakeholder from './components/PFCAdmin/AdminNewStakeholder';
+import CreateWastePicker from './components/DWCC/CreateWastePicker';
+import CreateExternalDWCC from './components/DWCC/CreateExternalDWCC';
 import './common.css';
 
 const store = createStore(rootReducer);
